+++ src/features/bin-designer/store/slices/index.ts
@@ -1,5 +1,5 @@
 export { createParamSlice } from './paramSlice';
-export { createCutoutSlice } from './cutoutSlice';
+export { createCutoutSlice, remainingCutoutCapacity } from './cutoutSlice';
 export { createHistorySlice } from './historySlice';
 export { createUISlice } from './uiSlice';
 export { createPersistenceSlice } from './persistenceSlice';